pyproject.toml
[project]
name = "django-template"
version = "0.1.0"
description = ""
authors = [
    {name = "Hailemichael Atrsaw T.",email = "hailemichael.atrsaw2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "markdown (>=3.8,<4.0)",
    "django-filter (>=25.1,<26.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
